refactor: centralised types under the `types` directory.
