+++ src/core/tool-calling/index.ts
@@ -5,4 +5,4 @@
 export { parseToolCallsFromContent, cleanContentFromToolCalls } from "./parser.js";
 export { executeToolCalls, executeToolCall } from "./executor.js";
 export { promptToolApproval } from "../../ui/input.js";
-export type { ToolExecutionResult } from "./executor.js";
+export type { ToolExecutionResult } from "../../types/index.js";
