+++ src/types/mcp.ts
@@ -1,0 +1,13 @@
+export interface MCPServer {
+  name: string;
+  command: string;
+  args?: string[];
+  env?: Record<string, string>;
+}
+
+export interface MCPTool {
+  name: string;
+  description?: string;
+  inputSchema?: any;
+  serverName: string;
+}
