+++ src/types/ui.ts
@@ -1,0 +1,11 @@
+export interface Colors {
+  white: string;
+  black: string;
+  primary: string;
+  tool: string;
+  secondary: string;
+  success: string;
+  error: string;
+  blue: string;
+  orange: string;
+}
